pyproject.toml
[project]
name = "trashpin"
version = "1.0.0"
description = "Discord bot that archives pinned messages to a designated channel"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "py-cord>=2.7.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
    "mypy>=1.19.1",
]
